pyproject.toml
[project]
name = "fastapi-portfolio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt==4.1.2",
    "fastapi>=0.122.0",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.6",
    "passlib>=1.7.4",
    "pydantic-settings>=2.12.0",
    "pydantic[email,settings]>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.44",
    "uvicorn[standard]>=0.38.0",
]
